NEW - bug 205: TLC stops if Queue Size grows to more than 2^31 http://bugzilla.tlaplus.net/show_bug.cgi?id=205 - convert len from int to long to support more than 2^31 states in state queue
git-svn-id: svn+ssh://svn.msr-inria.inria.fr/var/lib/svn/repository/tla/trunk/tla2-inria@23686 76a6fc44-f60b-0410-a9a8-e67b0e8fc65c

+++ tlatools/src/tlc2/tool/distributed/TLCServer.java
@@ -479,8 +479,8 @@ public class TLCServer extends UnicastRemoteObject implements TLCServerRMI,
 	/**
 	 * @return
 	 */
-	private int getNewStates() {
-		int res = stateQueue.size();
+	private long getNewStates() {
+		long res = stateQueue.size();
 		for (int i = 0; i < threadCnt; i++) {
 			res += threads[i].getCurrentSize();
 		}
@@ -527,7 +527,7 @@ public class TLCServer extends UnicastRemoteObject implements TLCServerRMI,
     {
         long statesGenerated = this.getStatesComputed();
 		long distinctStates = this.fpSetManager.size();
-		int statesLeftInQueue = this.getNewStates();
+		long statesLeftInQueue = this.getNewStates();
 		int level = this.trace.getLevelForReporting();
 		if (TLCGlobals.tool) {
             MP.printMessage(EC.TLC_PROGRESS_STATS, new String[] { String.valueOf(level),
