+++ tlatools/src/tlc2/tool/distributed/selector/StatisticalBlockSelector.java
@@ -24,7 +24,7 @@ public class StatisticalBlockSelector extends LimitingBlockSelector {
 	 *            The current size of the state queue.
 	 * @return The intended block size
 	 */
-	protected int getBlockSize(final int size, final TLCWorkerRMI aWorker) {
+	protected long getBlockSize(final int size, final TLCWorkerRMI aWorker) {
 		// has to be correct type and statistics have to be available
 		if(aWorker instanceof TLCWorkerSmartProxy) {
 			final TLCWorkerSmartProxy proxy = (TLCWorkerSmartProxy) aWorker;
